MSI-565: IsProductSalable conditions pool
