+++ app/code/Magento/InventoryConfiguration/Model/IsProductSalableCondition/ManageStockCondition.php
@@ -5,45 +5,60 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventoryCatalog\Model;
+namespace Magento\InventoryConfiguration\Model\IsProductSalable;
 
+use Magento\CatalogInventory\Api\StockConfigurationInterface;
 use Magento\InventoryConfigurationApi\Api\Data\StockItemConfigurationInterface;
 use Magento\InventoryConfigurationApi\Api\GetStockItemConfigurationInterface;
 use Magento\InventorySalesApi\Api\IsProductSalableInterface;
 
 /**
- * Class IsBackOrderAvailable
+ * @inheritdoc
  */
-class IsBackOrderAvailable implements IsProductSalableInterface
+class ManageStockCondition implements IsProductSalableInterface
 {
+    /**
+     * @var StockConfigurationInterface
+     */
+    private $configuration;
+
     /**
      * @var GetStockItemConfigurationInterface
      */
     private $getStockItemConfiguration;
 
     /**
+     * @param StockConfigurationInterface $configuration
      * @param GetStockItemConfigurationInterface $getStockItemConfiguration
      */
     public function __construct(
+        StockConfigurationInterface $configuration,
         GetStockItemConfigurationInterface $getStockItemConfiguration
     ) {
         $this->getStockItemConfiguration = $getStockItemConfiguration;
+        $this->configuration = $configuration;
     }
 
     /**
-     * @param string $sku
-     * @param int $stockId
-     * @return bool
+     * @inheritdoc
      */
     public function execute(string $sku, int $stockId): bool
     {
         /** @var StockItemConfigurationInterface $stockItemConfiguration */
         $stockItemConfiguration = $this->getStockItemConfiguration->execute($sku, $stockId);
-
-        if ($stockItemConfiguration->getBackorders() !== StockItemConfigurationInterface::BACKORDERS_NO) {
-            return true;
+        $globalManageStock = $this->configuration->getManageStock();
+        $manageStock = false;
+        if ((
+                $stockItemConfiguration->getUseConfigManageStock() == 1 &&
+                $globalManageStock == 1
+            ) || (
+                $stockItemConfiguration->getUseConfigManageStock() == 0 &&
+                $stockItemConfiguration->getManageStock() == 1
+            )
+        ) {
+            $manageStock = true;
         }
 
-        return false;
+        return !$manageStock;
     }
 }
